fix no device and no path isues

+++ main.py
@@ -20,34 +20,46 @@ def caputure():
     return ps.do_detect()
 
 def removeAllImages(ordner_pfad, tag):
-    for datei in os.listdir(ordner_pfad):
-        if (tag in datei) and (('.jpg' in datei) or ('.png' in datei)):
-            bild_pfad = os.path.join(ordner_pfad, datei)
-            os.remove(bild_pfad)
+    if os.path.exists(ordner_pfad):
+        for datei in os.listdir(ordner_pfad):
+            if (tag in datei) and (('.jpg' in datei) or ('.png' in datei)):
+                bild_pfad = os.path.join(ordner_pfad, datei)
+                os.remove(bild_pfad)
 
-if __name__ == "__main__":
-    ordner_pfad = os.getcwd()
-    
-    # removeAllCaptures(ordner_pfad,'capture')
-    removeAllImages(os.path.join(ordner_pfad,"processdImages"),'')
-    
-    startCapture()
-    image = caputure()
-    stopCapture()
-    
-    start_time = time.time()
-    
-    if not "image" in locals():
-        picture = os.path.join(ordner_pfad,"capture_1.png")
-        image = cv2.imread(picture, cv2.IMREAD_GRAYSCALE)
+def createImageFolder(ordner_pfad, tag):
+    if not os.path.exists(ordner_pfad):
+        os.mkdir(tag)
 
+def process(image):
     image = roi.main(image)
     cv2.imwrite( os.path.join(ordner_pfad,"processdImages","roi.jpg"),image)
     image = preProcess.main(image)
     cv2.imwrite( os.path.join(ordner_pfad,"processdImages","preProcess.jpg"),image)
     image = postProcess.skel(image)
     cv2.imwrite( os.path.join(ordner_pfad,"processdImages","postProcess.jpg"),image)
-    image = filters.main(postProcess.invert(image))
+    return  filters.main(postProcess.invert(image))
+    
+if __name__ == "__main__":
+    ordner_pfad = os.getcwd()
+    removeAllImages(os.path.join(ordner_pfad,"processdImages"),'')
+    createImageFolder(os.path.join(ordner_pfad,"processdImages"),'processdImages')
+
+    try:
+        startCapture()
+        image = caputure()
+        stopCapture()
+    except:
+        print("no device found, switch to simulation")
+    
+    start_time = time.time()
+    
+    if not "image" in locals():
+        picture = os.path.join(ordner_pfad,"capture_1.png")
+        image = cv2.imread(picture, cv2.IMREAD_GRAYSCALE)
+    else:
+         cv2.imwrite( os.path.join(ordner_pfad,"processdImages","capture.jpg"),image)
+
+    image = process(image)
     cv2.imwrite( os.path.join(ordner_pfad,"processdImages","filter.jpg"),image)
 
     end_time = time.time()
